updated drinks serialization to display recipes with ingredients

+++ db/migrate/20220323225411_create_recipes.rb
@@ -3,7 +3,7 @@ class CreateRecipes < ActiveRecord::Migration[6.1]
     create_table :recipes do |t|
       t.belongs_to :drink, null: false, foreign_key: true
       t.belongs_to :ingredient, null: false, foreign_key: true
-      t.integer :parts
+      t.float :parts
 
       t.timestamps
     end
